backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
python-dotenv==1.0.1
groq==0.9.0
langgraph==0.2.0
langchain-core==0.2.43
langchain-groq==0.1.9
pydantic==2.9.2
pydantic-settings==2.10.1
httpx==0.27.0
python-multipart==0.0.9
psycopg2-binary==2.9.9